AUI-1683 More tests for initial screen
Tests showed a bug where we were reusing the same Layout instance between different FormBuilder instances. We need to use valueFn instead of value for default attribute values.

Also noticed that we needed to update the empty message when the layout changed as well, instead of just when the layout row's changed.